% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFuns.R
\name{cievd_vector}
\alias{cievd_vector}
\title{Internal function to compute EVD quantile confidence intervals}
\usage{
cievd_vector(xrow, p, ci = 0.95)
}
\arguments{
\item{xrow}{A single-row data.frame of EVD parameters and covariance matrix}

\item{p}{Probability value (e.g. 0.99)}

\item{ci}{Confidence level (default 0.95)}
}
\value{
Numeric vector of confidence interval width(s)
}
\description{
Internal function to compute EVD quantile confidence intervals
}
\keyword{internal}
