% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFuns.R
\name{df_qevd}
\alias{df_qevd}
\title{Return a vector of EVD Quantiles}
\usage{
df_qevd(x, p, evd_mod_str, interval = NULL, lower.tail = TRUE, cores = 1)
}
\arguments{
\item{x}{A data.frame of EVD parameters}

\item{p}{Probability value (e.g. 0.99)}

\item{evd_mod_str}{One of "fgumbel", "fgev", or "fgumbelx"}

\item{interval}{Optional; interval passed to uniroot when needed}

\item{lower.tail}{Logical; if TRUE (default), computes P(\eqn{x \le y})}

\item{cores}{Number of cores to use for parallel processing (default is 1)}
}
\value{
A numeric vector of quantiles
}
\description{
Return a vector of EVD Quantiles
}
