% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFuns.R
\name{plot_emperical}
\alias{plot_emperical}
\title{Plot the Emperical Return Level Data}
\usage{
plot_emperical(x, xns = NULL, unitz = "-", ...)
}
\arguments{
\item{x}{A numeric vector, which may contain missing values.}

\item{xns}{A numeric vector, corrected for the non-stationary change in location, which may contain missing values.}

\item{unitz}{y-label}

\item{...}{parameters sent to base::plot}
}
\value{
r
}
\description{
Plot the Emperical Return Level Data
}
\examples{
ns = seq(-1,1,,50)
x = evd::rgev(50,loc=3)+ns
xns = x-ns
plot_emperical(x,xns)
}
