% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFuns.R
\name{qevd_vector}
\alias{qevd_vector}
\title{Return a Vector or EVD Quantiles}
\usage{
qevd_vector(x, p, evd_mod_str, interval = NULL)
}
\arguments{
\item{x}{vector of EVD paramters}

\item{p}{vector of probabilities.}

\item{evd_mod_str}{either a string "fgumbel", "fgev" or "fgumbelx" from the extreme value distribution (evd) in the evd package}

\item{interval}{A length two vector containing the end-points of the interval to be searched for the quantiles, passed to the uniroot function.}
}
\value{
gives the quantile function corresponding to p
}
\description{
Return a Vector or EVD Quantiles
}
\examples{
qevd_vector(c(1,0.5),0.5,"fgumbel")
}
